% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cci_score.R
\name{cci_score}
\alias{cci_score}
\title{Calculate Quan-weighted CCI score for mixed ICD9/ICD10 codes}
\usage{
cci_score(icd_data)
}
\arguments{
\item{icd_data}{A data.frame with id, code, and vocabulary_id variables}
}
\value{
data.frame with id, CCI constiuent phenotype indicators, and CCI score
}
\description{
Calculate Quan-weighted CCI score for mixed ICD9/ICD10 codes
}
\examples{
\dontrun{
cci_score(icd_data)
}
}
