% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{check_args}
\alias{check_args}
\title{Check argument types, length, or dimension}
\usage{
check_args(arg, type, length = NULL, dim = NULL)
}
\arguments{
\item{arg}{An argument to be checked.}

\item{type}{A character vector of candidate argument type.}

\item{length}{A numeric value of argument length or `NULL`.}

\item{dim}{A numeric vector of argument dimension or `NULL`.}
}
\value{
Check failure detailed error message.
}
\description{
Copied from Merck/simtrial R package
Accessed source code from https://rdrr.io/cran/simtrial/src/R/check_args.R on December 18, 2024.
}
\details{
If `type`, `length` or `dim` is `NULL`, the corresponding check will not be executed.
}
\section{Specification}{

\if{latex}{
 \itemize{
   \item Check if arg is NULL.
   \item Extract the type, length and dim information from arg.
   \item Compare with target values and report error message if it does not match.
 }
 }
\if{html}{The contents of this section are shown in PDF user manual only.}
}

\examples{
\dontrun{
tbl <- as.data.frame(matrix(1:9, nrow = 3))
check_args(arg = tbl, type = c("data.frame"))

vec <- c("a", "b", "c")
check_args(arg = vec, type = c("character"), length = 3)
}
}
\keyword{internal}
