% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox_helpers.R
\name{clean_coxph}
\alias{clean_coxph}
\title{Clean output from coxph object}
\usage{
clean_coxph(x)
}
\arguments{
\item{x}{A `coxph` object}
}
\value{
tibble reporting coefficient, se, HR (95% CI), and p-value
}
\description{
Clean output from coxph object
}
\examples{
\dontrun{
coxph_obj <- survival::coxph(...)
clean_coxph(coxph_obj)
}
}
