% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox_helpers.R
\name{clean_coxph_output}
\alias{clean_coxph_output}
\title{Clean output from coxph or coxphf object}
\usage{
clean_coxph_output(x)
}
\arguments{
\item{x}{A `coxph` or `coxphf` object}
}
\value{
tibble reporting coefficient, se, HR (95% CI), and p-value
}
\description{
Clean output from coxph or coxphf object
}
\examples{
\dontrun{
coxph_obj <- survival::coxph(...)
coxphf_obj <- coxphf::coxphf(...)
clean_coxph_output(coxph_obj)
clean_coxph_output(coxphf_obj)
}
}
