% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox_helpers.R
\name{clean_coxphf}
\alias{clean_coxphf}
\title{Clean output from coxphf object}
\usage{
clean_coxphf(x)
}
\arguments{
\item{x}{A `coxphf` object}
}
\value{
tibble reporting coefficient, se, HR (95% CI), and p-value
}
\description{
Clean output from coxphf object
}
\examples{
\dontrun{
coxphf_obj <- coxphf::coxphf(...)
clean_coxphf(coxphf_obj)
}
}
