% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox_helpers.R
\name{extract_cox}
\alias{extract_cox}
\title{Quickly run adjusted and unadjusted matched and unmatched Cox proportional hazards models}
\usage{
extract_cox(
  data,
  time,
  outcome,
  exposure,
  covariates = NULL,
  firth = FALSE,
  caliper = 0.1,
  match_std_mean_diff_threshold = 0.1
)
}
\arguments{
\item{data}{data.frame}

\item{time}{Character string corresponding to time variable}

\item{outcome}{Character string corresponding to outcome (status) variable}

\item{exposure}{Character string corresponding to exposure (treatment) variable}

\item{covariates}{Character vector corresponding to covariates; also serves as predictors in matching model}

\item{firth}{Logical; indicates whether to fit coxph (FALSE) or Firth bias-corrected coxph (`coxphf::coxphf`; TRUE)}

\item{caliper}{Caliper width for matching}

\item{match_std_mean_diff_threshold}{Standardized mean difference threshold; if std_mean_diff is greater than threshold, variable is included as a covariate in disease model in matched cohort}
}
\value{
A list object containing unadjusted unmatched Cox PH model (`unadjusted_unmatched`), unadjusted matched Cox PH model (unadjusted_matched), matching object (`match_obj`), adjusted matched Cox PH model (`adjusted_matched`), and a tibble summarizing coefficients from the models (`coef_sum`)
}
\description{
Quickly run adjusted and unadjusted matched and unmatched Cox proportional hazards models
}
\examples{
\dontrun{
patterns <- c("age", "sex", "race")
char_vec <- c("age", "sex", "raceWhite", "raceBlack")
matching_patterns(patterns, char_vec)
}
}
