% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox_helpers.R
\name{make_cox_formula}
\alias{make_cox_formula}
\title{Generate a Surv formula}
\usage{
make_cox_formula(time_var, status_var, x, covs = NULL, as_formula = FALSE)
}
\arguments{
\item{time_var}{Character name of time variable}

\item{status_var}{Character name of status variable}

\item{x}{Character name for exposure of interest variable}

\item{covs}{Character vector of covariate variables}

\item{as_formula}{Logical; whether to return formula as character (FALSE) or formula (TRUE)}
}
\value{
a character or formula object representing a Surv formula
}
\description{
Generate a Surv formula
}
\examples{
\dontrun{
time_var <- "time"
status_var <- "outcome"
exposure <- "glp1"
covariates <- c("age", "race")
make_cox_formula(
  time_var   = time_var,
  status_var = status_var,
  x          = exposure,
  covs       = covariates
)
}
}
