% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox_helpers.R
\name{matching_patterns}
\alias{matching_patterns}
\title{Helper function to extract variables from models or matching objects}
\usage{
matching_patterns(patterns, character_vector)
}
\arguments{
\item{patterns}{Character vector of patterns to explore and return}

\item{character_vector}{Character vector to check for presence of patterns}
}
\value{
A character vector of patterns present in `character_vector`
}
\description{
Helper function to extract variables from models or matching objects
}
\examples{
\dontrun{
patterns <- c("age", "sex", "race")
char_vec <- c("age", "sex", "raceWhite", "raceBlack")
matching_patterns(patterns, char_vec)
}
}
