% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{rmst_diff_plot}
\alias{rmst_diff_plot}
\title{Generate plot of difference in restricted mean survival time between groups across timepoints}
\usage{
rmst_diff_plot(
  matched_data,
  time_var = "time",
  outcome_var = "outcome",
  trt_var = "glp1",
  taus = 1:5
)
}
\arguments{
\item{matched_data}{dataset with treatment and event indicators and a time-to-event variable}

\item{time_var}{character name of the time-to-event variable, assumes days}

\item{outcome_var}{character name of the outcome (event) indicator variable}

\item{trt_var}{character name of the treatment indicator variable}

\item{taus}{vector of time points at which to assess restricted mean survival time, assumes years}
}
\value{
tibble reporting coefficient, se, HR (95% CI), and p-value
}
\description{
Generate plot of difference in restricted mean survival time between groups across timepoints
}
\examples{
\dontrun{
# example forthcoming
rmst_diff_plot(data)
}
}
