% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{rmst_plot}
\alias{rmst_plot}
\title{Generate restricted mean survival time plot}
\usage{
rmst_plot(
  x,
  trt_var = "glp1",
  tte_var = "time",
  event_var = "outcome",
  ref = "0",
  .tau = 365 * 5
)
}
\arguments{
\item{x}{dataset with treatment and event indicators and a time-to-event variable}

\item{trt_var}{character name of the treatment indicator variable}

\item{tte_var}{character name of the time-to-event variable}

\item{event_var}{character name of the event indicator variable}

\item{ref}{reference group of the treatment indicator variable}

\item{.tau}{time point at which to assess restricted mean survival time}
}
\value{
tibble reporting coefficient, se, HR (95% CI), and p-value
}
\description{
Generate restricted mean survival time plot
}
\examples{
\dontrun{
# example forthcoming
rmst_plot(data)
}
}
